alpha 1.1 bug fix

+++ app/src/main/java/com/example/fyp/menucreator/domain/foodCategory/DeleteFoodCategoryUseCase.kt
@@ -1,5 +1,7 @@
 package com.example.fyp.menucreator.domain.foodCategory
 
+import com.example.fyp.account_management.data.model.Account
+import com.example.fyp.account_management.data.model.AccountType
 import com.example.fyp.menucreator.data.repository.MenuSettingsRepository
 import com.example.fyp.menucreator.util.UiState
 import javax.inject.Inject
@@ -7,8 +9,11 @@ import javax.inject.Inject
 class DeleteFoodCategoryUseCase @Inject constructor(
     private val repo: MenuSettingsRepository
 ) {
-
-    suspend operator fun invoke(id: String, result: (UiState<Boolean>) -> Unit){
+    suspend operator fun invoke(account: Account, id: String, result: (UiState<Boolean>) -> Unit){
+        if (account.accountType !in listOf(AccountType.Manager, AccountType.Admin)){
+            result.invoke(UiState.Failure(Exception("Account has no permission!")))
+            return
+        }
         repo.deleteCategory(id,result)
     }
 }
